# Use AlternateLookup with ReadOnlySpan<char>